refactoring connection into a package
